Cоздал компоненту citiesBlock и вынес туда  состояние выбранного предложения
