+++ src/components/pages/main/main-page.tsx
@@ -5,6 +5,7 @@ import { TCity, TOffer, TOffers } from '../../../types/offers.ts';
 import OfferList from '../../offer-list/offer-list.tsx';
 import Map from '../../map/map.tsx';
 import SortForm from '../../sort-form/sort-form.tsx';
+import CitiesBlock from '../../cities-block/cities-block.tsx';
 
 
 type TMainPageProps = {
@@ -16,22 +17,11 @@ function MainPage({ offers, cities }: TMainPageProps): React.JSX.Element {
   const [selectedCity, setSelectedCity] = useState<TCity>(cities[0]);
   const [offersList, setOffersList] = useState<TOffers>(offers);
 
-  const [
-    selectedOffer,
-    setSelectedOffer
-  ] = useState<Pick<TOffer, 'id'> | undefined>(undefined);
-
   function handelSelectedCity(city: TCity) {
     const offersListByCity = offers.filter((offer) => offer.city.name === city.name);
 
-    setSelectedCity(city);
     setOffersList(offersListByCity);
-  }
-
-  function handleSelectedOffer(id: string) {
-    if (selectedOffer?.id !== id) {
-      setSelectedOffer({ id });
-    }
+    setSelectedCity(city);
   }
 
   useEffect(() => {
@@ -50,19 +40,7 @@ function MainPage({ offers, cities }: TMainPageProps): React.JSX.Element {
             <CitiesList onSelectedCity={handelSelectedCity} selectedCity={selectedCity}/>
           </section>
         </div>
-        <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offersList.length} places to stay in {selectedCity.name}</b>
-              <SortForm/>
-              <OfferList offers={offersList} onSelectedOffer={handleSelectedOffer} page={'main'}/>
-            </section>
-            <div className="cities__right-section">
-              <Map offers={offersList} selectedCity={selectedCity} selectedOffer={selectedOffer} page={'main'}/>
-            </div>
-          </div>
-        </div>
+        <CitiesBlock offersList={offersList} selectedCity={selectedCity}/>
       </main>
     </div>
   );
